chore: moving user folder to modules
