+++ app/modules/user/repository/UserRepository.ts
@@ -1,7 +1,7 @@
 import { EntityRepository, Repository } from 'typeorm'
 
-import hash_password from '../../services/becrypt/hashPassword'
-import User from '../../src/entities/User'
+import hash_password from '../../../services/becrypt/hashPassword'
+import User from '../../../src/entities/User'
 import { ICreateUser } from '../@types'
