+++ app/routers.ts
@@ -3,12 +3,12 @@ import { Router } from "express";
 import { isAuthenticated } from "./middlewares/IsAuthenticated";
 import { IsAdmin } from "./middlewares/IsAdmin";
 
-import userRouter from './user/routes'
+import userRouter from './modules/user/routes'
 import tokenJWTRouter from "./authenticate/router/AutheticateUserRouter";
 
 const routers = Router()
 
-routers.use('/user', isAuthenticated, IsAdmin, userRouter)
+routers.use('/user', userRouter)
 routers.use('/token', tokenJWTRouter)
 
 export default routers
